requirements.txt
keras==2.9.0
mediapipe==0.8.10.1
numpy==1.23.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
pandas==1.4.3
Pillow==8.4.0
tensorflow==2.9.1
tqdm==4.64.0